+++ android/expoview/src/main/java/host/exp/exponent/storage/ExperienceDBObject.java
@@ -5,14 +5,15 @@ package host.exp.exponent.storage;
 import com.raizlabs.android.dbflow.annotation.Column;
 import com.raizlabs.android.dbflow.annotation.PrimaryKey;
 import com.raizlabs.android.dbflow.annotation.Table;
+import com.raizlabs.android.dbflow.annotation.Unique;
 import com.raizlabs.android.dbflow.structure.BaseModel;
 
 @Table(database = ExponentDB.class)
 public class ExperienceDBObject extends BaseModel {
 
-  @Column
+  @Column(name = "id")
   @PrimaryKey
-  public String id;
+  public String scopeKey;
 
   @Column
   public String manifestUrl;