+++ android/versioned-abis/expoview-abi41_0_0/src/main/java/abi41_0_0/host/exp/exponent/modules/universal/ScopedFacebookModule.java
@@ -13,7 +13,7 @@ import abi41_0_0.org.unimodules.core.arguments.ReadableArguments;
 import abi41_0_0.org.unimodules.core.interfaces.LifecycleEventListener;
 
 import abi41_0_0.expo.modules.facebook.FacebookModule;
-import host.exp.exponent.kernel.ExperienceId;
+import host.exp.exponent.kernel.ExperienceKey;
 
 public class ScopedFacebookModule extends FacebookModule implements LifecycleEventListener {
   private final static String ERR_FACEBOOK_UNINITIALIZED = "ERR_FACEBOOK_UNINITIALIZED";