+++ android/versioned-abis/expoview-abi41_0_0/src/main/java/abi41_0_0/host/exp/exponent/modules/universal/notifications/ScopedNotificationsHandler.java
@@ -5,29 +5,29 @@ import android.content.Context;
 import abi41_0_0.expo.modules.notifications.notifications.handling.NotificationsHandler;
 import expo.modules.notifications.notifications.model.Notification;
 import expo.modules.notifications.notifications.model.NotificationResponse;
-import host.exp.exponent.kernel.ExperienceId;
+import host.exp.exponent.kernel.ExperienceKey;
 import host.exp.exponent.notifications.ScopedNotificationsUtils;
 
 public class ScopedNotificationsHandler extends NotificationsHandler {
-  private ExperienceId mExperienceId;
+  private ExperienceKey mExperienceKey;
   private ScopedNotificationsUtils mScopedNotificationsUtils;
 
-  public ScopedNotificationsHandler(Context context, ExperienceId experienceId) {
+  public ScopedNotificationsHandler(Context context, ExperienceKey experienceKey) {
     super(context);
-    mExperienceId = experienceId;
+    mExperienceKey = experienceKey;
     mScopedNotificationsUtils = new ScopedNotificationsUtils(context);
   }
 
   @Override
   public void onNotificationReceived(Notification notification) {
-    if (mScopedNotificationsUtils.shouldHandleNotification(notification, mExperienceId)) {
+    if (mScopedNotificationsUtils.shouldHandleNotification(notification, mExperienceKey)) {
       super.onNotificationReceived(notification);
     }
   }
 
   @Override
   public boolean onNotificationResponseReceived(NotificationResponse response) {
-    if (mScopedNotificationsUtils.shouldHandleNotification(response.getNotification(), mExperienceId)) {
+    if (mScopedNotificationsUtils.shouldHandleNotification(response.getNotification(), mExperienceKey)) {
       return super.onNotificationResponseReceived(response);
     }
     return false;