+++ docs/components/plugins/api/APIDataTypes.ts
@@ -146,8 +146,9 @@ export type TypeDeclarationContentData = {
 };
 
 export type TypeDeclarationData = {
+  name?: string;
   declaration?: TypeDeclarationContentData;
-  type: string;
+  type?: string;
   types: TypeDefinitionData[];
   typeArguments?: TypeDefinitionData[];
 };