+++ docs/components/plugins/api/APISectionTypes.tsx
@@ -146,6 +146,16 @@ const renderType = ({ name, comment, type }: TypeGeneralData): JSX.Element | und
         </div>
       );
     }
+  } else if (type.type === 'intrinsic') {
+    return (
+      <div key={`generic-type-definition-${name}`}>
+        <H3Code>
+          <InlineCode>{name}</InlineCode>
+        </H3Code>
+        <CommentTextBlock comment={comment} />
+        <ReactMarkdown renderers={mdRenderers}>{'__Type:__ `' + type.name + '`'}</ReactMarkdown>
+      </div>
+    );
   }
   return undefined;
 };