+++ packages/expo-notifications/src/BackgroundNotificationTasksModule.native.ts
@@ -1,0 +1,5 @@
+import { NativeModulesProxy } from '@unimodules/core';
+
+import { BackgroundNotificationTasksModule } from './BackgroundNotificationTasksModule.types';
+
+export default NativeModulesProxy.ExpoBackgroundNotificationTasksModule as BackgroundNotificationTasksModule;